requirements.txt
# NovoExpert-2 dependencies (CatBoost + MapLight + GIN pipeline)

# Core ML
torch==2.5.0
catboost>=1.2.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Data
pandas>=1.5,<2.0
numpy>=1.24,<2.0
PyTDC>=0.4.0

# Cheminformatics
rdkit-pypi>=2023.9.1

# GIN embeddings (DGL-Life)
torchdata==0.7.1
dgl>=2.0  # Install from https://data.dgl.ai/wheels/ if pip default fails
dgllife>=0.3.2
